Move awx_plugins/tests/test_inventory.py to awx_plugins/tests/test_inventory.py
